Alpha Release 0.1
Supports three wave types: Sine, Square, and Triangle. No command-line arguments are being parsed yet, so these modes can be changed on Line 128, in the first argument of play_sound(). The baseline note is locked to A2 (110Hz), but can be edited on line 132, in the baseParam assignment. The message can be edited on the following line with the message variable.